consolidate commands
